+++ MicrosoftGraphAPIBot/Program.cs
@@ -52,6 +52,8 @@ namespace MicrosoftGraphAPIBot
                     services.AddScoped<GraphApi, OutlookApi>();
                     services.AddScoped<GraphApi, OneDriveApi>();
                     services.AddScoped<GraphApi, PermissionsApi>();
+                    services.AddScoped<GraphApi, CalendarApi>();
+                    services.AddScoped<GraphApi, PersonalContactsApi>();
                     services.AddTransient<ApiController>();
                     services.AddTransient<ApiCallManager>();
                     services.AddScoped<DefaultGraphApi>();