+++ ReleaseNotes/Changelog.md
@@ -3,6 +3,7 @@
 - 支援 OneDrive API [(#9)](https://github.com/NTUT-SELab/MicrosoftGraphBot/issues/9)
 - 支援 Permissions API [(#11)](https://github.com/NTUT-SELab/MicrosoftGraphBot/issues/11)
 - 支援 Calendar API [(#12)](https://github.com/NTUT-SELab/MicrosoftGraphBot/issues/12)
+- 支援 Personal contacts API [(#10)](https://github.com/NTUT-SELab/MicrosoftGraphBot/issues/10)
 
 ## 變更: